Rename files to singular "semaphore".
