+++ mock/semaphore.h
@@ -1,5 +1,5 @@
-#ifndef SEMAPHORES_MOCK_H_
-#define SEMAPHORES_MOCK_H_
+#ifndef SEMAPHORE_MOCK_H_
+#define SEMAPHORE_MOCK_H_
 
 /** A semaphore structure. The implementation is public to allow to inspect
  * state when in tests. */
